iso_test.go
package veil_api_client_go

import (
	"github.com/stretchr/testify/assert"
	"testing"
)

func Test_IsoListGet(t *testing.T) {
	client := NewClient("", "", false)
	response, _, err := client.Iso.List()
	assert.Nil(t, err)
	for _, v := range response.Results {
		entity, _, err := client.Iso.Get(v.Id)
		assert.Nil(t, err)
		assert.NotEqual(t, entity.Id, "", "Iso Id can not be empty")

		break
	}

	return

}

func Test_IsoUpload(t *testing.T) {
	t.SkipNow()
	client := NewClient("", "", false)
	response, _, err := client.DataPool.List()
	assert.Nil(t, err)
	if len(response.Results) == 0 {
		t.SkipNow()
	}
	firstDp := response.Results[0]
	iso, _, err := client.Iso.Create(firstDp.Id, "test_live.iso")
	assert.Nil(t, err)
	assert.NotEqual(t, iso.Id, "", "Iso Id can not be empty")

	return

}
